+++ src/adapters/SinonServer.spec.ts
@@ -1,9 +1,9 @@
 import sinon, { type SinonFakeXMLHttpRequest } from 'sinon';
 
 import { SinonServer } from './SinonServer.js';
-import { Single } from './Single.js';
-import { Collection } from './Collection.js';
-import type { BaseResponse } from './BaseServer.js';
+import { Single } from '../Single.js';
+import { Collection } from '../Collection.js';
+import type { BaseResponse } from '../AbstractBaseServer.js';
 
 function getFakeXMLHTTPRequest(
     method: string,
